requirements-sample.txt
python-dotenv>=1.0.0
Pillow>=9.5.0
requests
